+++ core/02-environment.php
@@ -66,3 +66,8 @@ $paths->upload_file_prefix = "Files/";
 // Create the cache directory if it doesn't exist
 if(!is_dir($paths->cache_directory))
 	mkdir($paths->cache_directory, 0700);
+
+// Set the user agent string
+$php_version = ini_get("expose_php") == "1" ? "PHP/".phpversion() : "PHP";
+ini_set("user_agent", "$php_version (".PHP_SAPI."; ".PHP_OS." ".php_uname("m")."; ".(PHP_INT_SIZE*8)." bits; rv:$version) Pepperminty-Wiki/$version-".substr($commit, 0, 7));
+unset($php_version);